Simplify build scripts

+++ .run.ps1
@@ -1,6 +1,13 @@
-& .git-version.ps1
-Write-Host "Opening http://localhost:8088/swagger/index.html"
+& .init.ps1
+
+dotnet publish -nologo --configuration Release --runtime linux-x64 --no-self-contained -p:OutputType=Exe -p:DebugOutput=None src\Terragate.sln
+
+if ($LASTEXITCODE -ne 0) {
+    return
+}
+
 Start-Process "chrome.exe" "http://localhost:8088/swagger/index.html" -ErrorAction SilentlyContinue
 
-Write-Host "Executing docker run..."
+docker build -t $global:currentTag .
 docker run -p 8088:80 -it -e ASPNETCORE_ENVIRONMENT=Development -e TF_VAR_VRA_USER=$env:TF_VAR_VRA_USER -e TF_VAR_VRA_PASS=$env:TF_VAR_VRA_PASS --rm $global:currentTag
+
